cloud-functions/rec-retrieve-suggestions/requirements.txt
functions-framework==3.*
requests
google-cloud-storage
sentence_transformers
numpy
pandas
scikit-learn